pyproject.toml
[project]
name = "K8sResourceResizer"
version = "0.2.0"
description = "A tool to manage k8s pod resources"
readme = "README.md"
license = "MIT"
requires-python = ">=3.13"
dependencies = [
    "click>=8.3.1",
    "gitpython>=3.1.46",
    "loguru>=0.7.3",
    "numpy>=2.4.3",
    "pandas>=3.0.1",
    "pmdarima>=2.1.1",
    "prophet>=1.3.0",
    "pygithub>=2.8.1",
    "python-dotenv>=1.2.2",
    "pyyaml>=6.0.3",
    "requests>=2.32.5",
    "scikit-learn>=1.8.0",
    "statsmodels>=0.14.6",
]

[tool.setuptools]
packages = ["src"]
package-dir = { "src" = "src" }

[build-system]
requires = ["setuptools>=61", "wheel", "cython>=0.29.21,<3.0", "numpy>=1.21.2"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "commitizen>=4.13.9",
    "prek>=0.3.5",
]

[tool.ruff]
line-length = 88
